Change Scopus query to get all available fields
Full results can only be requested from subscriber IP addresses.
We don't have to worry about this for deployment on Carolina CloudApps
(since the domain is already under unc.edu),
but for queries to work locally we need to set the HTTP header
'X-Forwarded-For' to our remote IP address.
For now at least I'm just putting this address in a secret .key file.